Update patch set 6
Patch Set 6: Looks good to me, but someone else must approve

ifneq is just inefficient logic when is most cases it won't be used. Choosing to use an open-ended variable instead of a true/false meant switching the order, which was a last minute decision to save excess declarations. It has been corrected.

Patch-set: 6