Update patch set 2
Patch Set 2:

Why not fix it once in one place instead of individually for each device tree?

The problem is if the file is in the wrong format, brcm_patchram_plus doesn't fail, you just get a default MAC address, which will generally work for most things, and the maintainer doesn't notice.  I've noticed it on three devices so far, the latest being the p5wifi (with is based on p4 which also doesn't have the bdaddr_read hack so I'm guessing it also doesn't get the right bluetooth MAC either).

Patch-set: 2